requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-community
sentence-transformers
faiss-cpu
groq
python-dotenv